pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-summarizer"
version = "0.1.0"
description = "A tool for summarizing documents and code using AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "oha", email = "aaronoh2015@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=0.19.0",
    "pykomodo>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
]

[project.scripts]
docsummarizer = "document_summarizer.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["document_summarizer"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"